% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliaPlotR-package.R
\docType{package}
\name{ReliaPlotR-package}
\alias{ReliaPlotR}
\alias{ReliaPlotR-package}
\title{ReliaPlotR: Interactive Reliability Probability Plots}
\description{
Build interactive Weibull Probability Plots with 'WeibullR' by David Silkworth and Jurgen Symynck (2022) \url{https://CRAN.R-project.org/package=WeibullR}, an R package for Weibull analysis, and 'plotly' by Carson Sievert (2020) \url{https://plotly-r.com}, an interactive web-based graphing library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulgovan.github.io/ReliaPlotR/}
  \item \url{https://github.com/paulgovan/ReliaPlotR}
  \item Report bugs at \url{https://github.com/paulgovan/ReliaPlotR/issues}
}

}
\author{
\strong{Maintainer}: Paul Govan \email{paul.govan2@gmail.com} (\href{https://orcid.org/0000-0002-1821-8492}{ORCID}) [copyright holder]

}
\keyword{internal}
